// src/types.ts
// This is now the single source of truth for what an Employee is.
export interface Employee {
    id: string;
    code: string;
    name: string;
    date: string;
    designation: string;
    department: string;
    location: string;
    payslip: string;
    gender: string;
    status: string;
  }